Fix Massive logic, timestamp type, add test
Fix multiple bugs in Massive: rename/initialize lastRequest, null-check and reset requestCount correctly, use MAX_REQUESTS constant, format month/day with leading zeros, fix URL path segment, print URL, and update lastRequest assignment. Change AggregateBars.timestamp from int to long to avoid overflow. Add MassiveTest to exercise getAggregate (loads API key from test resource). Update UserTest to capture and log the generated password length before asserting the exception. These changes improve correctness and add coverage.

+++ src/main/java/com/example/api/massive/records/AggregateBars.java
@@ -18,7 +18,7 @@ public record AggregateBars(
         @JsonProperty("c") double closePrice,
         @JsonProperty("h") double highestPrice,
         @JsonProperty("l") double lowestPrice,
-        @JsonProperty("t") int timestamp,
+        @JsonProperty("t") long timestamp,
         @JsonProperty("n") int transactions
     ) {}
 }
